requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
neo4j
python-dotenv
